Modified the Exercise to Support Undo/Redo.
This was more challenging than expected due to how I implemented it in the previous exercise, which did not require deep copy of objects.  This exercise did require deep copy of objects, which seemed ill-fit considering the scope.  Instead, Undo/Redo now captures Actions that can be played to affect a desired state.  This puts some cross-cutting concerns on the application that can perhaps be abstracted away better, but at least the present implementation works without need for deep copy.